update app class and create separete class for menu
